@@ -18,12 +18,12 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
-  path_provider: 1.6.18
-  http: 0.12.1
-  universal_html: 1.2.3
+  path_provider: 2.0.2
+  http: 0.13.3
+  universal_html: 2.0.8
   flutter:
     sdk: flutter
 
